feat : add aiservice, message, room

+++ src/room/room.service.ts
@@ -1,21 +1,28 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { RoomRepository } from './room.repository';
-import { CreateRoomDto, EditRoomDto } from './dto/room.dto';
+import { CreateRoomDto, EditRoomDto } from './dto/create-room.dto';
 
 @Injectable()
 export class RoomService {
   constructor(private readonly repo: RoomRepository) {}
 
-  async getAll() {
-    return await this.repo.getAllRooms();
+  async getAll(id: string) {
+    const rooms = await this.repo.getAllRooms(id);
+    console.log(rooms);
+    return rooms;
   }
 
   async getRoomById(id: string) {
+    const roomExist = await this.repo.getRoomById(id);
+    if (!roomExist) throw new NotFoundException(`r`);
     return await this.repo.getRoomById(id);
   }
 
   async deleteRoomById(id: string) {
+    const isRoomExist = await this.repo.getRoomById(id);
+    if (isRoomExist == null || isRoomExist == undefined)
+      throw new NotFoundException(`room with id : ${id} not found`);
     return await this.repo.deleteRoom(id);
   }
 
@@ -23,7 +30,11 @@ export class RoomService {
     return await this.repo.editRoom(data, id);
   }
 
-  async createRoom(data: CreateRoomDto) {
-    return await this.repo.createRoom(data);
+  async createRoom(data: CreateRoomDto, id: string) {
+    const room = await this.repo.createRoom(data, id);
+    return {
+      id: room.id,
+      name: room.name,
+    };
   }
 }
